refactor: move _pdg module to particle
